doc for avg func
